+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2026-04-12
+
+### Added
+- `clearHistory<T>()` to clear stored events without disabling collection
+- `disableHistory<T>()` to stop collecting and discard stored events
+
 ## [0.3.0] - 2026-04-05
 
 ### Added